Complete `Person` init method
Notice that I've also updated the defaults, so people are assumed to be "not vaccinated".

This allows us to generate large numbers of people and then pick a few to vaccinate or infect after, rather than having to know exactly which random ID's to vaccinate or infect. Makes filters and such easier. (For example: determining a set of n users who are not vaccinated, so we can infect them.)

Also adjusted some method docstrings.