Reconfigure `Simulation` class
Now, more logic is within the `cycle_step` method. This does make the function larger, but I've seperated out with comments to make it clearer.

I've done this so I don't have to toss around parameters like its nobody's business ;) plus it's just cleaner-looking code.

You'll also find that the print method is a bit more comprehensive, and was moved to the `Logger` class.